fix a tag query test
we've always defined that there can only be one expression with a
tag-operator per query, no more than that. but the old behavior was that
it simply ignore one of those expressions when two were specified, the
new behavior is that the query instantiation returns a validation error.
so the test needs to be adjusted accordingly.